requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Cache & Queue
redis>=5.0.0
celery[redis]>=5.3.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pyotp>=2.9.0

# Storage
boto3>=1.34.0

# Vector DB
qdrant-client>=1.7.0

# ML/NLP
httpx>=0.26.0
sentence-transformers>=2.3.0
tiktoken>=0.6.0
openai>=1.12.0

# Document parsing
pdfplumber>=0.10.0
python-docx>=1.1.0

# Testing
pytest-timeout>=2.2.0
